Fix compile issues

+++ RingCentral/RestClient.cs
@@ -50,7 +50,7 @@ namespace RingCentral
                 {
                     if (!refreshScheduled)
                     { // don't do duplicate schedule
-                        Task.Delay((_token.expires_in.Value - 120) * 1000).ContinueWith((action) =>
+                        Task.Delay((int)(_token.expires_in.Value - 120) * 1000).ContinueWith((action) =>
                         { // 2 minutes before expiration
                             refreshScheduled = false;
                             Refresh();
